Extract MetricKotlinPoet
